# static/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: NW Monitor API
  description: |
    REST API for the Network Monitor System.
    Provides endpoints for device management, monitoring, topology,
    dashboards, alerts, SLA reports, bandwidth, SNMP traps, syslog,
    user management, audit logs, and system administration.
  version: "1.2.0"
  contact:
    name: NW Monitor Admin
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/dashboards:
    get:
      tags: [Dashboards]
      summary: List all dashboards
      responses:
        "200":
          description: Array of dashboards
    post:
      tags: [Dashboards]
      summary: Create a dashboard (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                layout_config: { type: object }
                is_public: { type: integer, default: 0 }
      responses:
        "200":
          description: Dashboard created
  /api/dashboards/{dashboard_id}:
    get:
      tags: [Dashboards]
      summary: Get a specific dashboard
      parameters:
        - $ref: "#/components/parameters/DashboardId"
      responses:
        "200":
          description: Dashboard detail
    put:
      tags: [Dashboards]
      summary: Update a dashboard (admin)
      parameters:
        - $ref: "#/components/parameters/DashboardId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                layout_config: { type: object }
                description: { type: string }
                is_public: { type: integer }
      responses:
        "200":
          description: Dashboard updated
    delete:
      tags: [Dashboards]
      summary: Delete a dashboard (admin)
      parameters:
        - $ref: "#/components/parameters/DashboardId"
      responses:
        "200":
          description: Dashboard deleted
  /api/dashboards/reorder:
    post:
      tags: [Dashboards]
      summary: Reorder dashboards (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dashboard_ids]
              properties:
                dashboard_ids:
                  type: array
                  items: { type: integer }
      responses:
        "200":
          description: Order saved

  # ---- Alerts ----
...
  /api/bandwidth/interfaces:
    get:
      tags: [Bandwidth]
      summary: Get interfaces for a device
      parameters:
        - name: device_id
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Interface list
  /api/bandwidth/history:
    get:
      tags: [Bandwidth]
      summary: Get bandwidth time-series data
      parameters:
        - name: device_id
          in: query
          required: true
          schema: { type: integer }
        - name: if_index
          in: query
          schema: { type: integer }
        - name: minutes
          in: query
          schema: { type: integer, default: 60 }
      responses:
        "200":
          description: Bandwidth history
...
  /api/bandwidth/poll:
    post:
      tags: [Bandwidth]
      summary: Trigger manual bandwidth poll for a device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id: { type: integer }
      responses:
        "200":
          description: Poll started
...
  /api/bandwidth/debug/{device_id}:
    get:
      tags: [Bandwidth]
      summary: SNMP diagnostic for a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Diagnostic results

  # ---- SNMP Traps ----
...
components:
  parameters:
    DeviceId:
      name: device_id
      in: path
      required: true
      schema: { type: integer }
    DashboardId:
      name: dashboard_id
      in: path
      required: true
      schema: { type: integer }
...